requirements.txt
asgiref==3.5.2
Django==4.1
django-service-objects==0.7.1
djangorestframework==3.13.1
Pillow==9.2.0
product==0.1.1.1
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
